structure clean-up
